requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
google-cloud-storage==2.18.2
google-cloud-documentai==2.28.0
google-cloud-aiplatform==1.70.0
requests==2.32.3
